Remove not used code
