--- pkg/tsdb/cloudwatch/get_metric_data_test.go
@@ -1,13 +1,0 @@
-package cloudwatch
-
-import (
-	"testing"
-
-	. "github.com/smartystreets/goconvey/convey"
-)
-
-func TestCloudWatchGetMetricData(t *testing.T) {
-	Convey("CloudWatchGetMetricData", t, func() {
-
-	})
-}
